backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# API & Serialization
pydantic==2.5.3

# Observability
structlog==24.1.0
prometheus-client==0.19.0
django-prometheus==2.3.1
python-json-logger==2.0.7

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment
python-dotenv==1.0.0

# Code Quality (Development)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
bandit==1.7.6

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0
